#104 auth refactor
