+++ laravel/routes/api.php
@@ -20,8 +20,8 @@ Route::get('sortList',  function() {
 | is assigned the "api" middleware group. Enjoy building your API!
 |
 */
-Route::get('login', 'Api\UserController@login');
-Route::resource('users', 'Api\UserController', ['only' => ['store']]);
+Route::post('login', 'Api\AuthController@login');
+Route::resource('users', 'Api\AuthController', ['only' => ['store']]);
 Route::resource('games', 'Api\GameController', ['only' => ['index', 'show']]);
 Route::resource('combos', 'Api\ComboController', ['only' => ['index', 'create', 'show', 'update', 'destroy']]);
 Route::resource('moves', 'Api\MoveController', ['only' => ['index']]);
